refactor: use SlotMap

+++ gui/src/wrapper/project/timeline.h
@@ -32,7 +32,7 @@ class Timeline {
     Layer layerByLayerHandle(size_t layer_handle) const noexcept;
     Layer layerSortedAt(uint32_t index) const noexcept;
 
-    std::tuple<Clip, size_t> findClipById(uint64_t id) const noexcept;
+    std::tuple<Clip, uint32_t> findClipById(uint64_t id) const noexcept;
     bool canPlaceClipAt(uint32_t layerIdx, int64_t position, int64_t duration,
                         const std::set<uint64_t> &exclude_set) const;
